feat: add support for Redux DevTools [WT-7]
Allows Redux DevTools to be used for debugging.
