+++ package.json
@@ -44,6 +44,8 @@
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.3.0",
     "redux": "4.2.0",
+    "redux-devtools-extension": "^2.13.9",
+    "redux-thunk": "^2.4.1",
     "simplebar-react": "^2.3.7",
     "validator": "^13.7.0"
   },
