+++ src/store.js
@@ -1,17 +1,8 @@
-import { createStore } from 'redux'
+import { createStore, applyMiddleware } from 'redux'
+import { composeWithDevTools } from 'redux-devtools-extension'
+import rootReducer from '../src/reducers/index'
+import thunk from 'redux-thunk'
 
-const initialState = {
-  sidebarShow: true,
-}
+const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)))
 
-const changeState = (state = initialState, { type, ...rest }) => {
-  switch (type) {
-    case 'set':
-      return { ...state, ...rest }
-    default:
-      return state
-  }
-}
-
-const store = createStore(changeState)
 export default store
